refactor:  UI와 관련된 건 모두 View로 바라보도록 수정
